refactor(aws-creds): move to aws-creds folder
