+++ src/aws-creds/aws-creds.module.ts
