+++ src/file/file.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { FileController } from './file.controller';
 import { FileService } from './file.service';
 import { ConfigModule } from '../config/config.module';
-import { AWSCredsModule } from '../auth/aws-creds.module';
+import { AWSCredsModule } from '../aws-creds/aws-creds.module';
 
 @Module({
   imports: [ConfigModule, AWSCredsModule],
